MVVM
Arrumei na arquitetura MVVM
