+++ CloudKitTest/Preview Content/Assets.xcassets/AccentColor.colorset/Contents.json
