+++ CloudKitTest/Preview Content/Assets.xcassets/AppIcon.appiconset/Contents.json
