+++ CloudKitTest/Preview Content/Assets.xcassets/Contents.json
